clock: freescale bsp uses ipg as a root clock for the perclk
there may be another clock (per_podf) fix to add here. BSP clock handling
is fucked up to say the least at linux-legacy levels and later BSPs really
get out of hand so it is hard to say what is being done or not. But there
is definitely evidence of using the ipg clock (666MHz ->) as a distant
i2c2 relative and not the PLL PODF.