sgtl5000: don't prematurely switch to internal LDO
On probe SGTL5000 codec can use an optional VDDD supply or an internal LDO
to generate it's 1.2V core voltage. In the case of an external VDDD the
probe logic would bulk get all regulators and then, on SUCCESS, replace
VDDD with the LDO.

This defeats later logic where a version check is in place (for chip id
0x11 or higher) to use the internal LDO to work around an errata, since
external_vddd is never set, it never runs through this replacement of
the regulators.

In theory, now, on older hardware the external VDDD supply will be used
and on newer hardware (0x11 or above as previously stated) it will boot,
use external VDDD, check the version register, then swap out for the LDO
only if needed, UNLESS VDDD is missing in which case it will try for the
internal LDO anyway (although the ability for this regulator to get
enabled if any other supply was missing anyway is debatable...)