clk: fix cpu0 clock, don't fiddle with parent PLL rate for the CPU clock
We can reparent the ARM root clock away from PLL1 to change PLL1's rate but
with the current clock subsystem none of this happens. By using the
imx_clk_divider helper we would be setting SET_RATE_PARENT flag. I can only
assume that implies strange things like trying to make the parent rate
different so it can get a more accurate clock for the children.

And if it *is* doing so, it would be very hard to see it in action since it
would lock the CPU. And I think it IS locking the CPU.

Introduce imx_clk_divider_fixedparent which makes no such implication
to the clk API, such that the divider is just fudged to a close enough
value and the parent would remain unchanged.

That seems to have cleared up some kind of random lockups (random as in
I can't control CPU usage as finely as I'd need to get empirical data)

Also fix the cpu_podf clock so it's called cpu0 so that the cpufreq driver
can find it.