mc13892: sanity check on the number of regulators parsed vs. registered
Imagine a situation where a device tree has a few regulators in an
appropriate node:

regulators {
	sw1 {
		..
	};

	vvideo {
		..
	};

	vfake {
		..
	};

	vfake2 {
		..
	};
};

In the above example, the node name "vfake" is an attempt to match a
regulator name inside the driver which just so happens to not exist.

In this case, the number of regulators the mc13892 driver thinks it has
does not match the number of regulators it parsed and registered. Since
it will go over this array based on this number, it will actually
re-register regulator "0" (which happens to be SW1) over and over
again until it reaches the number.

So what you get is this at the end of your regulator kernel messages:

SW1: at 1100 mV
SW1: at 1100 mV

.. up to that number. You can also add trash nodes inside your regulator
node (like vfake { .. };) for the same effect.

This patch effectively sanity checks the number of counted children of
the regulators node vs. the number that actually matched driver names,
and sets the appropriate num_regulators value. It also gives a little
warning for device tree authors that they MAY have screwed something up.

This patch will break mc13783 for now.

There probably needs to be a documentation update such that there is a
warning NOT to add non-relevant data into the regulators node?