clock-hacking: move fixed clocks and pll1-3 to real DT matching
This simplifies the clk-imx51-imx53 stuff by not having to register
fixed clocks and pass their rates around for hacky registration.

Any clk[table_id] lookups for those clocks had to be converted to
a __clk_lookup("name").

Little bit extra code and this can be put in drivers/clk/ instead
of the architecture folder.