gpt: i.MX3,5,6 GPT clock rework, now DT "pure", seperate sched_clock init
This essentially aligns with Highbank and anything with an architected/twd
timer (Cortex-A9) whereby timer initialization (i.e. clocksource/clockevent)
is somewhat seperated from the setup_sched_clock call.

This way multiple timers can be registered for the platform, and the more
timers the better for the system (many high resolution timers firing
rarely compared to one high resolution timer constantly serving events
should equal lower power consumption) and the sched_clock can be set up
on the more stable of those timers (on i.MX the architected timer would
be preferred - so i.MX6 guys I wonder why you're not using it - but on
i.MX5 this could be GPT or EPIT depending on use of the "other" timer
type).

This screws non-DT platforms as they have no way of initializing the
timer anymore, but a small stub (or resurrection of mxc_timer_init)
and a little split of the register function could bring it back.

Caveat: on i.MX53 and i.MX6 there are two EPIT timers. How the hell
do we register TWO timers as above and then pick ONE of them as a
possible sched_clock source? If we convert EPIT to the same DT-pure
framework like the architected or twd...?