sched_clock: use DT to define linux,scheduler-clock property
Without any timer with this property, it will use the jiffies clock
(32-bit counter, 100Hz) for the scheduler which is godawful but it
DOES work.

TODO: make a workalike of_clk_init (of_timers_init?) which matches
all timer nodes we want to find, and automagically register the
scheduler clock. First, we make EPIT like GPT code and test it as
the scheduler.