clocksource: lots of changes, mostly adding EPIT timer
EPIT ported to the "device tree pure" method as per GPT.

GPT is still sched_clock.

EPIT is now registered as an alternative clocksource but since it
counts backwards it seems unsuitable for sched_clock (how did MX35
use it? CONFIG_MXC_USE_EPIT is confusing and probably ancient legacy
where it was activated but nothing touched it, and MX35 used jiffies
instead. I wonder how nobody noticed..)

Add some debug here and there.

Rename our timers to timer@ since gpt/epit are not real node names.
The aliases give the reality of it all.

EPIT is not well tested yet but since it's not "in use" for anything
that can be done later.