gpt: make the GPT the delay timer
BogoMIPS will go down to 133.0 since it's an exact double of the timer
frequency. Please ignore this, it's not a problem (in fact it should
be more accurate as the LPJ calibration is skipped)

Not sure if EPIT can be the delay timer as it counts down and the
architected timer implementation seems to want an unsigned long.. so
what we do is we use GPT for delay timer (since 0->UINT_MAX is fine,
it will just need resetting) and EPIT for sched_clock (since sched_clock
handles it). As long as nobody wants to delay for 64 seconds (GPT
wrap time) it is all fine. We will run into a huge problem if the GPT
clock rate is much, much faster however..