clocks: move clk-imx51-imx53 to drivers/clk/clk-imx5-ccm
Move it into the clocks framework like it should be. Actually starting
to think I should have put it in the imx/ directory like mxs is?

In the process of moving; remove hardcoded addresses in favor of offsets
to the of_iomapped base register.

Reorder the mx5_clocks_common_init so that imx_clk_mux calls are at the
top and further order them by register order and bit order so it looks
like the docs (somewhat). This makes it look like clk-imx6q which is
much, much more readable (albeit only with a monitor capable of displaying
120 characters wide..)

Tested: uart, audio, usb (otg at least), pata, mmc, display (di0, hdmi)
and all seem to be working fine, clock tree seems reasonable and a
sampling of rates (/sys/kernel/debug/clk/ --> /clk_rate) shows identical
to linux-legacy or at least very close to it (clock frameworks are a
little different so it's not possible to get the precise same rates),
delay_timer and sched_clock working as expected, still.

Now, to make this clumsy mess into something clean by moving those
functions where it prints clocks and stuff into the platform code and
not the clock management code..