clk-imx5-ccm: remove clock fixups from the core init code
Except one. I am not sure about this per_root -> ipg parenting, it seems
that the GPT and EPIT functionality can use a high or low resolution
clock (maybe 133MHz and 66MHz? Or 66MHz and 33MHz?) but this might only
be reasonable on devices like i.MX35. On i.MX51 and 53 becasue of this
hack both of them get set to 66MHz and we set a bit in the GPT at least
to run off the.. "high frequency" IPG clock (which is this reparent).

Anyway I left it in there until I figure out where the crap it goes.