dt, sgtl5000: clean up the mclk node by removing unused gpio property
.. but keep it around in a new node for a clock which doesn't exist in the
Linux kernel yet (since there's a couple hundred lines of code needed to
be written to manage it properly). Give it a crazy compatible so that it
doesn't ever get probed for now. This node is NOT to be published!!!

Also clean up the eSDHC GPIO comments (no point repeating the same info)
while we're here