smartbook: correct device tree regulators etc.
Fix spi frequency to match Smarttop (20MHz might be a bit high)

Fix regulator names - note, the same tree from Smarttop applies so
we should share this (including always-on and boot-on crap). Remove
VCAM. Dunno about RTC supply for SB, add a comment for now.

Add regulator rtc property so we get a time on Smartbook instead of
going back to 1970.