XXX drm/amdgpu: re-enable suspend phase 2 for S0ix
This really needs to be done to properly tear down
the device.

For testing:
I added some commented out code to skip IP blocks in
that function.  Enable that code and try commenting
out different combinations of IP blocks to see what
is problematic if the code does not work as is.

Signed-off-by: Alex Deucher <alexander.deucher@amd.com>