arm64: dts: qcom: c630: Fix soundcard setup
Currently Soundcard has 1 rx device for headset and SoundWire Speaker Playback.

This setup has issues, ex if we try to play on headset the audio stream is
also sent to SoundWire Speakers and we will hear sound in both headsets and speakers.

Make a seperate device for Speakers and Headset so that the streams are
different and handled properly.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>